Add files via upload
Flight ticket prices can be something hard to guess, today we might see a price, check out the price of the same flight tomorrow, it will be a different story. We might have often heard travellers saying that flight ticket prices are so unpredictable. Huh! Here we take on the challenge! As data scientists, we are gonna prove that given the right data anything can be predicted. Here you will be provided with prices of flight tickets for various airlines between the months of March and June of 2019 and between various cities.

As data scientists, we were supposed to prove that given the right data anything can be predicted. Here we have been provided with prices of flight tickets for various airlines between the months of March and June of 2019 and between various cities.
With the observation of departure time, arrival time, duration of the flight, route taken by the flight and the name of the airline I able to predict 97% of correct prediction with F1 score metrics.